Much further.
There are still no edges yet, but our nodes exist in 3d spacetime and can be crudely filtered and updated.

Next job is to build edges. Edges should refer to an existing node, and should be applied as a strengethening of nodes.

That means, an edge exists in spacetime

Edges will provide the source input value to the destination node across the entire life of the edge.